requirements.txt
requests==2.27.1
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.4
matplotlib==3.8.4
statsmodels==0.14.2
scikit-learn==1.4.2
seaborn== 0.12.2
statsmodels==0.14.2